% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(n, x, y)
}
\arguments{
\item{n}{number of bootstrap replicates}

\item{x}{variable}

\item{y}{variable}
}
\value{
Returns the correlation between x and y, the bootstrap estimate of correlation, standard error, bias and confidence interval.
}
\description{
Simulates Buffon's experiment of throwing needles on a striped floor
}
\author{
Maria Frey & Cecilie Neupart
}
\keyword{bias}
\keyword{bootstrap}
\keyword{confidence}
\keyword{correlation}
\keyword{error}
\keyword{estimate}
\keyword{interval}
\keyword{standard}

