% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof}
\alias{gof}
\title{Goodness of fit}
\usage{
gof(x, p=rep(1/length(x),length(x)))
}
\arguments{
\item{x}{vector of observed data}

\item{p}{expected probability (default uniform)}
}
\value{
Returns the chi-square goodness of fit value.
}
\description{
Uses observed and expected data to calculate the goodness of fit by the chi-square test
}
\author{
Maria Frey & Cecilie Neupart
}
\keyword{chi}
\keyword{fit}
\keyword{goodness}
\keyword{of}
\keyword{square}

