% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_chisq.R
\name{MC_chisq}
\alias{MC_chisq}
\title{Monte Carlo Chi squared distribution}
\usage{
MC_chisq(x, df, n)
}
\arguments{
\item{x}{variable}

\item{df}{degrees of freedom}

\item{n}{number of random numbers}
}
\value{
Returns the p-value, i.e. P(chi^2 > x)
}
\description{
Applies Monte Carlo to generate a chi-squared distribution with \code{d}f degrees of freedom by generating \code{n} numbers for each variable.
}
\author{
Maria Frey & Cecilie Neupart
}
\keyword{Monte}
\keyword{carlo}
\keyword{chi}
\keyword{degrees}
\keyword{distribution}
\keyword{freedom}
\keyword{of}
\keyword{p-value}
\keyword{square}

